requirements.txt
# python 3.8
python-telegram-bot[socks]==12.4.1
topicaxis-opengraph==0.5
